refactor: Abstract cart rows into CartProduct component
